Create README.md
The MAXREFDES89# provides an ideal interface for anyone developing with brushed DC motors. The design is an mbed®-compatible, Arduino® form factor shield for the rapid development of brushed DC motor applications. The shield contains four MAX14871 full-bridge DC motor drivers for driving up to 4 motors. A MAX17501 DC-DC converter allows the system to operate from a single 7 to 36VDC power supply. Four MAX4704 4:1 multiplexers permit setting the current regulation mode of the MAX14871, while two MAX5387 digital potentiometers provide the ability to set the motor current limit. A MAX7300 GPIO expander supports interfacing each motor driver circuit to the mbed-enabled microcontroller platform.

An mbed library for the shield is available at the following link, MAX14871_Shield Library. The library abstracts the details of interfacing with the supporting hardware of the reference design and allows the user to focus on their application.

The MAXREFDES89# is a tool for rapid development that offers optimal flexibility to the user during development of their application. In a final design, the MAX14871 may be the only required silicon for driving a brushed DC motor and regulating the motor’s current.